Licenses: Clarify exceptions for standalone apps
On 2010-01-27, an email [1] was sent to the mailing list by Wolfgang
Denk which clarified the intended licensing exceptions for standalone
applications. As the "export.h" header and the "stubs.c" source files
are required to implement a standalone application, the intention was
that these files be covered by the licensing exception. This is made
clear in the following quotes from that email:

	"exports.h" should be added to the "allowed" file list; there should
	be no need to include "common.h". Eventually this needs fixing.
	Patches are welcome.

	"examples/standalone/stubs.c" should be added to the "allowed" file
	list (the ppc_*jmp.S files are LGPLed).

	There should be no doubts - the intention is clear, the current state
	may need improvement. Help (read: patches) welcome.

[1]: https://lists.denx.de/pipermail/u-boot/2010-January/067174.html

Signed-off-by: Paul Barker <paul.barker@sancloud.com>
Cc: Wolfgang Denk <wd@denx.de>
Acked-by: Wolfgang Denk <wd@denx.de>